pull up r19891 from trunk
 r19891@cathode-dark-space:  jaltman | 2007-08-28 16:58:45 -0400
 ticket: new
 subject: NIM: 64-bit Windows Support and Removal of Compile Time Warnings
 component: windows
 
 This patch permits Network Identity Manager to be built for 64-bit Windows.
 In the process all compile time warnings have been taken care of.  
 
 For 64-bit Windows, we do not build the Kerberos v4 Credential Provider
 and we will not attempt to load the krb524 library.  
 
 Note that when testing the 64-bit NIM, there is no CCAPI at the 
 moment so you must manually specify a FILE: ccache as part of the 
 identity's Kerberos v5 configuration if you want to use cache's 
 other than the MSLSA.
 
 This patch also consolidates the computation of the default ccache 
 name into utility functions:
 
   khm_krb5_get_identity_default_ccache
   khm_krb5_get_identity_default_ccacheA
 
 

ticket: 5696

git-svn-id: svn://anonsvn.mit.edu/krb5/branches/krb5-1-6@20005 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/windows/identity/kmm/kmm.c
@@ -61,7 +61,7 @@ khm_boolean kmmint_load_locale_lib(kmm_module_i * m, kmm_module_locale * l)
 
         EnterCriticalSection(&cs_kmm);
         m->h_resource = h;
-        m->lcid_resource = l->language;
+        m->lcid_resource = (WORD) l->language;
         LeaveCriticalSection(&cs_kmm);
 
         return TRUE;
@@ -72,7 +72,7 @@ khm_boolean kmmint_load_locale_lib(kmm_module_i * m, kmm_module_locale * l)
 
         EnterCriticalSection(&cs_kmm);
         m->h_resource = m->h_module;
-        m->lcid_resource = l->language;
+        m->lcid_resource = (WORD) l->language;
         LeaveCriticalSection(&cs_kmm);
 
         return TRUE;
